+++ packages/ui-app/src/Headers/types.ts
@@ -1,18 +1,7 @@
 // Copyright 2026 @polkadot-cloud/polkadot-cloud-apps authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { ExtensionInjected } from '@w3ux/types'
 import type { ComponentType } from 'react'
-
-declare global {
-	interface Window {
-		walletExtension?: {
-			isNovaWallet?: boolean
-		}
-		injectedWeb3?: Record<string, ExtensionInjected>
-		opera?: boolean
-	}
-}
 
 export interface HeadersProps {
 	Sync: ComponentType
