+++ packages/ui-app/src/env.d.ts
@@ -1,0 +1,14 @@
+// Copyright 2026 @polkadot-cloud/polkadot-cloud-apps authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+import type { ExtensionInjected } from '@w3ux/types'
+
+declare global {
+	interface Window {
+		walletExtension?: {
+			isNovaWallet?: boolean
+		}
+		injectedWeb3?: Record<string, ExtensionInjected>
+		opera?: boolean
+	}
+}
